1-webrtc/requirements.txt
# Run on python 3.11
pipecat-ai==0.0.76
pipecat-ai[webrtc,openai,noisereduce,daily,groq,google]
pipecat-ai-small-webrtc-prebuilt==0.7.0
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
aiohttp==3.11.18
requests==2.32.4
loguru==0.7.3
onnxruntime==1.22.1